Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/html/semantics/interactive-elements/the-menu-element/menuitem-label-expected.txt

item2 item 3 item4

FAIL Menuitem.label should be the label attribute content or the result of stripping and collapsing ASCII whitespace from the child text content. assert_equals: expected (string) "item1" but got (undefined) undefined
